Add intermediary AbstractTileMenu class
