+++ src/main/java/com/verdantartifice/primalmagick/common/tiles/base/TileInventoryPM.java
@@ -158,14 +158,6 @@ public class TileInventoryPM extends TilePM /* implements WorldlyContainer */ {
         }
     }
 
-    public boolean stillValid(Player player) {
-        if (this.level.getBlockEntity(this.worldPosition) != this) {
-            return false;
-        } else {
-            return player.distanceToSqr((double)this.worldPosition.getX() + 0.5D, (double)this.worldPosition.getY() + 0.5D, (double)this.worldPosition.getZ() + 0.5D) <= 64.0D;
-        }
-    }
-
 //    @Override
 //    public void clearContent() {
 //        this.items.clear();
